MP3 Convertor
This project consists of code to convert a wav file to mp3.  I am using
Xuggler (http://www.xuggle.com/xuggler/) to perform the conversion.

The code I have written is a wrapper that will use Xuggler for
conversion.

When we pass in the byte stream of a wav file, the code will return the
mp3 converted ByteArrayOutputStream.